don't publicly export, move to _util
