+++ src/trio/_channel.py
@@ -5,7 +5,6 @@ from math import inf
 from typing import (
     TYPE_CHECKING,
     Generic,
-    NoReturn,
 )
 
 import attrs
@@ -21,11 +20,6 @@ if TYPE_CHECKING:
     from types import TracebackType
 
     from typing_extensions import Self
-
-import sys
-
-if sys.version_info < (3, 11):
-    from exceptiongroup import BaseExceptionGroup
 
 
 def _open_memory_channel(
@@ -446,63 +440,3 @@ class MemoryReceiveChannel(ReceiveChannel[ReceiveType], metaclass=NoPublicConstr
         See `MemoryReceiveChannel.close`."""
         self.close()
         await trio.lowlevel.checkpoint()
-
-
-def _raise(exc: BaseException) -> NoReturn:
-    """This helper allows re-raising an exception without __context__ being set."""
-    # cause does not need special handling, we simply avoid using `raise .. from ..`
-    __tracebackhide__ = True
-    context = exc.__context__
-    try:
-        raise exc
-    finally:
-        exc.__context__ = context
-        del exc, context
-
-
-# idk where to define this. It's a util, but exported, so _util doesn`t fit.
-# it'll be used by bg_with_channel, but is not directly related to channels.
-def raise_single_exception_from_group(
-    eg: BaseExceptionGroup[BaseException],
-) -> NoReturn:
-    """This function takes an exception group that is assumed to have at most
-    one non-cancelled exception, which it reraises as a standalone exception.
-
-    If a :exc:`KeyboardInterrupt` is encountered, a new KeyboardInterrupt is immediately
-    raised with the entire group as cause.
-
-    If the group only contains :exc:`Cancelled` it reraises the first one encountered.
-
-    It will retain context and cause of the contained exception, and entirely discard
-    the cause/context of the group(s).
-
-    If multiple non-cancelled exceptions are encountered, it raises
-    :exc:`AssertionError`.
-    """
-    cancelled_exceptions = []
-    noncancelled_exceptions = []
-
-    # subgroup/split retains excgroup structure, so we need to manually traverse
-    def _parse_excg(e: BaseException) -> None:
-        if isinstance(e, KeyboardInterrupt):
-            # immediately bail out
-            raise KeyboardInterrupt from eg
-
-        if isinstance(e, trio.Cancelled):
-            cancelled_exceptions.append(e)
-        elif isinstance(e, BaseExceptionGroup):
-            for sub_e in e.exceptions:
-                _parse_excg(sub_e)
-        else:
-            noncancelled_exceptions.append(e)
-
-    _parse_excg(eg)
-
-    if len(noncancelled_exceptions) > 1:
-        raise AssertionError(
-            "Attempted to unwrap exceptiongroup with multiple non-cancelled exceptions. This is often caused by a bug in the caller."
-        ) from eg
-    if len(noncancelled_exceptions) == 1:
-        _raise(noncancelled_exceptions[0])
-    assert cancelled_exceptions, "internal error"
-    _raise(cancelled_exceptions[0])
